ball processing logic added

+++ src/main/java/com/dugout/dugoutcore/dto/BallDto.java
@@ -20,7 +20,7 @@ public class BallDto extends BaseDto {
   UserDTO bowler;
   UserDTO wicketKeeper;
   Boolean isValid;
-  Long runs;
+  Integer runs;
   BallType type;
   Integer ballNumber;
   Integer batsmanRuns;
